refactor(router): move streaming routes to blueprint
