+++ app/routes.py
@@ -1840,504 +1840,6 @@ def init_routes(app: Flask) -> None:
                 400,
             )
 
-    @app.route("/stream.png")
-    @login_required
-    def stream_png():
-
-        group = request.args.get("group")
-        camera = request.args.get("camera")
-        if group == "all":
-            group = None
-        if camera == "all":
-            camera = None
-
-        if camera:
-            camera = validate_template_name(camera)
-            if camera is None:
-                abort(400, "Invalid camera name")
-            cam_path = os.path.join(
-                os.path.dirname(os.path.abspath(__file__)),
-                "..",
-                SCREENSHOT_DIRECTORY,
-                camera,
-                "latest_camera.png",
-            )
-            if os.path.exists(cam_path) and screenshots._is_valid_png(cam_path):
-                return send_conditional_file(cam_path, PNG_TTL_SEC)
-            abort(404)
-
-        if group:
-            group = secure_filename(group)
-            group_path = os.path.join(
-                os.path.dirname(os.path.abspath(__file__)),
-                "..",
-                SCREENSHOT_DIRECTORY,
-                f"{group}_latest_camera.png",
-            )
-            if os.path.exists(group_path) and screenshots._is_valid_png(group_path):
-                return send_conditional_file(group_path, PNG_TTL_SEC)
-            abort(404)
-
-        latest_path = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)),
-            "..",
-            SCREENSHOT_DIRECTORY,
-            "latest_camera.png",
-        )
-        if os.path.exists(latest_path):
-            if screenshots._is_valid_png(latest_path):
-                return send_conditional_file(latest_path, PNG_TTL_SEC)
-            logging.warning("Invalid latest camera image removed: %s", latest_path)
-            try:
-                os.remove(latest_path)
-            except OSError as exc:
-                logging.warning(
-                    "Failed to remove invalid latest image %s: %s",
-                    latest_path,
-                    exc,
-                )
-
-        global last_time, last_shot
-        # implement some simple caching so the server doesn't get crushed
-        if (
-            last_time
-            and time.time() - last_time < 1
-            and last_shot
-            and os.path.exists(last_shot)
-        ):
-            return send_conditional_file(last_shot, PNG_TTL_SEC)
-
-        templates = template_manager.get_templates()
-        sorted_templates = sorted(
-            templates.items(),
-            key=lambda x: (x[1].get("last_video_time", 0) or 0),
-            reverse=True,
-        )
-
-        most_recent_time = 0
-        most_recent_file = None
-        last_file = None
-        for _, template in sorted_templates:
-            name = validate_template_name(template.get("name"))
-            if name is None:
-                continue
-            path = os.path.join(
-                os.path.dirname(os.path.abspath(__file__)),
-                "..",
-                SCREENSHOT_DIRECTORY,
-                name,
-            )
-            files_with_mtime = []
-            for f in glob.glob(path + "/*.png"):
-                if not os.path.isfile(f) or f.endswith(".tmp.png"):
-                    # Ignore in-progress screenshots
-                    continue
-                try:
-                    mtime = os.path.getmtime(f)
-                except OSError:
-                    # File might have been removed between glob and stat
-                    continue
-                if not screenshots._is_valid_png(f):
-                    continue
-                files_with_mtime.append((f, mtime))
-            if not files_with_mtime:
-                continue
-            files_with_mtime.sort(key=lambda t: t[1])
-            last_file, last_mtime = files_with_mtime[-1]
-            if last_mtime > most_recent_time:
-                most_recent_file = last_file
-                most_recent_time = last_mtime
-        if most_recent_file is None:
-            if last_shot and os.path.exists(last_shot):
-                # Fall back to the previously served screenshot
-                return send_conditional_file(last_shot, PNG_TTL_SEC)
-            abort(404)
-
-        last_time = time.time()
-        last_shot = most_recent_file
-
-        if os.path.exists(most_recent_file) and screenshots._is_valid_png(
-            most_recent_file
-        ):
-            return send_conditional_file(most_recent_file, PNG_TTL_SEC)
-        if (
-            last_file
-            and os.path.exists(last_file)
-            and screenshots._is_valid_png(last_file)
-        ):
-            last_shot = last_file
-            return send_conditional_file(last_file, PNG_TTL_SEC)  # better than nothing
-        if (
-            last_shot
-            and os.path.exists(last_shot)
-            and screenshots._is_valid_png(last_shot)
-        ):
-            return send_conditional_file(last_shot, PNG_TTL_SEC)
-        abort(404)
-
-    @app.route(
-        "/test.rtsp",
-        methods=[
-            "OPTIONS",
-            "DESCRIBE",
-            "SETUP",
-            "PLAY",
-            "PAUSE",
-            "GET_PARAMETER",
-            "TEARDOWN",
-        ],
-    )
-    @login_required
-    def handle_rtsp():
-
-        session_id = request.headers.get("Session", str(uuid.uuid4()))
-        cseq = request.headers.get("CSeq", "0")
-
-        if request.method == "OPTIONS":
-            return Response(
-                "Public: OPTIONS, DESCRIBE, SETUP, PLAY, PAUSE, GET_PARAMETER, TEARDOWN",
-                headers={"CSeq": cseq},
-            )
-
-        elif request.method == "DESCRIBE":
-            sdp = (
-                "v=0\r\n"
-                "o=- 0 0 IN IP4 127.0.0.1\r\n"
-                "s=Glimpser RTSP Stream\r\n"
-                "t=0 0\r\n"
-                "m=video 0 RTP/AVP 26\r\n"
-                "a=rtpmap:26 JPEG/90000\r\n"
-                "a=control:streamid=0\r\n"
-            )
-            return Response(
-                sdp,
-                mimetype="application/sdp",
-                headers={"CSeq": cseq, "Content-Base": request.url},
-            )
-
-        elif request.method == "SETUP":
-            if session_id not in rtsp_sessions:
-                rtsp_sessions[session_id] = {
-                    "state": "READY",
-                    "seq": random.randint(0, 65535),
-                    "timestamp": random.randint(0, 0xFFFFFFFF),
-                    "ssrc": random.randint(0, 0xFFFFFFFF),
-                    "last_keepalive": time.time(),
-                }
-
-            transport = request.headers.get("Transport", "")
-            client_ports = (0, 0)
-            match = re.search(r"client_port=(\d+)(?:-(\d+))?", transport)
-            if match:
-                first = int(match.group(1))
-                second = int(match.group(2) or first + 1)
-                client_ports = (first, second)
-
-            server_ports = (5004, 5005)
-            session = rtsp_sessions[session_id]
-            session["client_ports"] = client_ports
-            session["server_ports"] = server_ports
-
-            transport_response = transport
-            if transport_response and not transport_response.endswith(";"):
-                transport_response += ";"
-            transport_response += f"server_port={server_ports[0]}-{server_ports[1]};ssrc={session['ssrc']}"
-
-            return Response(
-                headers={
-                    "CSeq": cseq,
-                    "Session": session_id,
-                    "Transport": transport_response,
-                }
-            )
-
-        elif request.method == "PLAY":
-            if session_id not in rtsp_sessions:
-                abort(454)  # Session Not Found
-            rtsp_sessions[session_id]["state"] = "PLAYING"
-            return Response(
-                headers={
-                    "CSeq": cseq,
-                    "Session": session_id,
-                    "RTP-Info": "url=rtsp://example.com/test.rtsp/streamid=0;seq=0;rtptime=0",
-                }
-            )
-
-        elif request.method == "PAUSE":
-            if session_id not in rtsp_sessions:
-                abort(454)
-            rtsp_sessions[session_id]["state"] = "PAUSED"
-            return Response(headers={"CSeq": cseq, "Session": session_id})
-
-        elif request.method == "GET_PARAMETER":
-            if session_id not in rtsp_sessions:
-                abort(454)
-            rtsp_sessions[session_id]["last_keepalive"] = time.time()
-            return Response(headers={"CSeq": cseq, "Session": session_id})
-
-        elif request.method == "TEARDOWN":
-            if session_id in rtsp_sessions:
-                del rtsp_sessions[session_id]
-            return Response(
-                headers={
-                    "CSeq": cseq,
-                    "Session": session_id,
-                }
-            )
-
-        return "Method Not Allowed", 405
-
-    @app.route("/motion.mjpg", methods=["GET"])
-    @login_required
-    def motion_mjpg():
-        group = request.args.get("group")
-        camera = request.args.get("camera")
-        if group == "all":
-            group = None
-        if camera == "all":
-            camera = None
-        return Response(
-            generate(group=group, camera=camera, filename="last_motion.png"),
-            mimetype="multipart/x-mixed-replace; boundary=frame",
-        )
-
-    @app.route("/caption.mjpg", methods=["GET"])
-    @login_required
-    def caption_mjpg():
-        group = request.args.get("group")
-        camera = request.args.get("camera")
-        if group == "all":
-            group = None
-        if camera == "all":
-            camera = None
-        logging.debug("last caption")
-        return Response(
-            generate(group=group, camera=camera, filename="last_caption.png"),
-            mimetype="multipart/x-mixed-replace; boundary=frame",
-        )
-
-    @app.route("/internal_caption.mjpg", methods=["GET"])
-    @login_required
-    def internal_caption_mjpg():
-        group = request.args.get("group")
-        camera = request.args.get("camera")
-        return Response(
-            generate_caption_loop(group=group, camera=camera),
-            mimetype="multipart/x-mixed-replace; boundary=frame",
-        )
-
-    @app.route("/motion_caption.mjpg", methods=["GET"])
-    @login_required
-    def motion_caption_mjpg():
-        group = request.args.get("group")
-        camera = request.args.get("camera")
-        if group == "all":
-            group = None
-        if camera == "all":
-            camera = None
-        logging.debug("last motion caption")
-        return Response(
-            generate(group=group, camera=camera, filename="last_motion_caption.png"),
-            mimetype="multipart/x-mixed-replace; boundary=frame",
-        )
-
-    @app.route("/test_pattern.mjpg", methods=["GET"])
-    @login_required
-    def test_pattern_mjpg():
-        spinner_frames = "⠋⠙⠹⠸⠼⠴⠦⠧⠇⠏"
-        index = 0
-
-        def generate_pattern():
-            nonlocal index
-            boundary = b"frame"
-            while True:
-                spinner = spinner_frames[index % len(spinner_frames)]
-                img = test_pattern.generate_test_pattern(spinner=spinner)
-                frame = test_pattern.encode_jpeg_with_metadata(img)
-                yield b"--" + boundary + b"\r\n"
-                yield b"Content-Type: image/jpeg\r\n\r\n" + frame + b"\r\n"
-                index += 1
-                time.sleep(1.0)
-
-        return Response(
-            generate_pattern(),
-            mimetype="multipart/x-mixed-replace; boundary=frame",
-        )
-
-    @app.route("/fast_stream.mjpg", methods=["GET"])
-    @login_required
-    def fast_stream_mjpg():
-        camera = request.args.get("camera")
-        if not camera:
-            abort(400, "camera parameter required")
-        if not template_manager.get_template(camera):
-            abort(404)
-        return Response(
-            generate_fast_mjpg(camera),
-            mimetype="multipart/x-mixed-replace; boundary=frame",
-        )
-
-    @app.route("/rtsp_stream")
-    def rtsp_stream():
-        session_id = request.args.get("session")
-        if (
-            session_id not in rtsp_sessions
-            or rtsp_sessions[session_id]["state"] != "PLAYING"
-        ):
-            abort(400, "Invalid session or session not in PLAYING state")
-        return Response(
-            generate(rtsp=True, session_id=session_id),
-            mimetype="application/x-rtp",
-        )
-
-    @app.route("/stream.mp4")
-    @login_required
-    def stream_mp4():
-        """Stream the latest MP4 for a camera or group."""
-
-        camera = request.args.get("camera")
-        if camera:
-            camera = validate_template_name(camera)
-            if camera is None:
-                abort(400, "Invalid camera name")
-            video_path = os.path.join(
-                os.path.dirname(os.path.abspath(__file__)),
-                "..",
-                VIDEO_DIRECTORY,
-                camera,
-                "in_process.mp4",
-            )
-            if not os.path.exists(video_path):
-                abort(404)
-            return Response(
-                stream_with_context(generate_video_stream(video_path)),
-                mimetype="video/mp4",
-            )
-
-        # Default group
-        lgroup = "all"
-
-        # Get the group from the request arguments and validate
-        group = request.args.get("group")
-        if group and re.match(r"^[a-zA-Z0-9_]+$", group):
-            lgroup = group
-        else:
-            # If the group is provided but invalid, return a 400 Bad Request
-            if group:
-                abort(400, "Invalid group name. Group name must be alphanumeric.")
-
-        lgroup = secure_filename(lgroup)
-        video_path = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)),
-            "..",
-            VIDEO_DIRECTORY,
-            f"{lgroup}_in_process.mp4",
-        )
-
-        if not os.path.exists(video_path):
-            abort(404)
-
-        # Stream the video in small chunks for continuous playback
-        return Response(
-            stream_with_context(generate_video_stream(video_path)),
-            mimetype="video/mp4",
-        )
-
-    @app.route("/live_video")
-    @login_required
-    def live_video():
-        camera = request.args.get("camera")
-        if not camera:
-            abort(400, "camera parameter required")
-        details = template_manager.get_template(camera)
-        if not details:
-            abort(404)
-        url = details.get("url")
-        if not url:
-            abort(404)
-        return Response(
-            stream_with_context(generate_live_stream(url)),
-            mimetype="video/mp4",
-        )
-
-    @app.route("/stream.m3u8")
-    @login_required
-    def playlist_m3u8():
-        path = os.path.join(
-            os.path.dirname(os.path.join(__file__)), "..", VIDEO_DIRECTORY
-        )
-
-        # Check if the video directory exists
-        if not os.path.exists(path):
-            abort(404)
-
-        camera = request.args.get("camera")
-        group = request.args.get("group")
-
-        # Generate playlist content
-        playlist_content = "#EXTM3U\n"
-        playlist_content += "#EXT-X-VERSION:3\n"
-        playlist_content += (
-            "#EXT-X-TARGETDURATION:10\n"  # Assuming each segment is up to 10 seconds
-        )
-        playlist_content += "#EXT-X-MEDIA-SEQUENCE:0\n"
-
-        templates = template_manager.get_templates()
-
-        filtered_templates = []
-        if camera:
-            camera = validate_template_name(camera)
-            if camera is None:
-                abort(400, "Invalid camera name")
-            details = templates.get(camera)
-            if details:
-                filtered_templates.append((camera, details))
-        elif group:
-            if not re.match(r"^[a-zA-Z0-9_]+$", group):
-                abort(400, "Invalid group name. Group name must be alphanumeric.")
-            for name, details in templates.items():
-                groups = [g.strip() for g in details.get("groups", "").split(",")]
-                if group in groups:
-                    valid = validate_template_name(name)
-                    if valid:
-                        filtered_templates.append((valid, details))
-        else:
-            for name, details in templates.items():
-                valid = validate_template_name(name)
-                if valid:
-                    filtered_templates.append((valid, details))
-
-        # Sort templates by 'last_video_time' descending
-        sorted_templates = sorted(
-            filtered_templates,
-            key=lambda x: (x[1].get("last_video_time", 0) or 0),
-            reverse=True,
-        )
-
-        for camera_name, _ in sorted_templates:
-            lkey = generate_timed_hash()
-            video_path = f"{request.url_root}last_video/{camera_name}?timed_key={lkey}"
-            playlist_content += f"#EXTINF:10.0,{camera_name}\n{video_path}\n"
-
-        playlist_content += "#EXT-X-ENDLIST\n"
-
-        return Response(playlist_content, mimetype="application/x-mpegURL")
-
-    @app.route("/stream")
-    @login_required
-    def stream():
-        # Get a list of active cameras (with updates within the last 1 day)
-        return render_template("stream.html", page_title="Stream")
-
-    @app.route("/groups")
-    @login_required
-    def get_groups():
-        # Assuming you have a function that returns a list of unique groups
-        groups = get_active_groups()
-        if "all" not in groups:
-            groups = ["all"] + groups
-        return jsonify(groups)
-
     @app.route("/captions")
     @login_required
     def captions():
